moving instance into a new file
